Move caliper from experiment to lib
