+++ experiments/babelstream/experiment.py
@@ -5,7 +5,7 @@
 
 from benchpark.directives import variant
 from benchpark.experiment import Experiment
-from benchpark.expr.builtin.caliper import Caliper
+from benchpark.caliper import Caliper
 from benchpark.cuda import CudaExperiment
 from benchpark.rocm import ROCmExperiment
 from benchpark.openmp import OpenMPExperiment
